Migrate widgets/player's content to use pure iced
Unfortunately we can't yet make player and character
component yet, because then we would lose the ability
to query their state. That will change with the next
version of iced, but that one removes the impure APIs
so we should switch to pure, keep the current widget
struct approach and refactor down the road.